requirements.txt
# Alexandria dependencies
fastapi==0.115.*
uvicorn[standard]==0.30.*
httpx==0.27.*
python-frontmatter==1.1.*
tqdm==4.*
pydantic==2.*
psutil==5.*
